Refactor scripts to src.
