+++ __init__.py
@@ -1,1 +1,17 @@
-from .scripts import *
+from .src import (
+    schemas,
+    load_entities,
+    establish_db_connection,
+    load_alias_entity_prior_probabilities,
+    parse,
+    namespaces,
+)
+
+__all__ = [
+    "schemas",
+    "load_entities",
+    "establish_db_connection",
+    "load_alias_entity_prior_probabilities",
+    "parse",
+    "namespaces",
+]
