+++ src/__init__.py
@@ -1,21 +1,17 @@
-from .wiki import (
+from .extraction import (
     schemas,
     load_entities,
     establish_db_connection,
-    extract_demo_dump,
     load_alias_entity_prior_probabilities,
     parse,
     namespaces,
 )
-from .utils import read_filter_terms
 
 __all__ = [
     "schemas",
     "load_entities",
     "establish_db_connection",
-    "extract_demo_dump",
     "load_alias_entity_prior_probabilities",
     "parse",
     "namespaces",
-    "read_filter_terms",
 ]
